fixing types convertions.
and removed Base64Encoder references.

+++ src/main/java/com/breathe/service/DeviceService.java
@@ -6,7 +6,7 @@ import com.breathe.model.DeviceModel;
  * Created by denis on 03.04.15.
  */
 public interface DeviceService {
-    public boolean addDevice(String deviceId, String deviceName, Integer delay, double co2MinLevel);
+    public boolean addDevice(String deviceId, String deviceName, int delay, int co2MinLevel);
 
     DeviceModel findDevicesById(String deviceId);
 }
